Move elevator transport map to transport manager.
