+++ src/game/ObjectMgr.cpp
@@ -2185,7 +2185,7 @@ void ObjectMgr::LoadGameobjects(bool reload)
         }
 
         if (gInfo->type == GAMEOBJECT_TYPE_TRANSPORT)
-            sMapMgr.AddElevatorTransportForMap(data.position.mapId, guid);
+            sTransportMgr.AddElevatorTransportForMap(data.position.mapId, guid);
         // if not this is to be managed by GameEvent System or Pool system
         else if (!alreadyPresent && gameEvent == 0 && GuidPoolId == 0 && EntryPoolId == 0)
             AddGameobjectToGrid(guid, &data);
