+++ src/game/Objects/GameObject.cpp
@@ -1046,7 +1046,7 @@ bool GameObject::IsTransport() const
     // If something is marked as a transport, don't transmit an out of range packet for it.
     GameObjectInfo const* gInfo = GetGOInfo();
     if (!gInfo) return false;
-    return gInfo->type == GAMEOBJECT_TYPE_TRANSPORT || gInfo->type == GAMEOBJECT_TYPE_MO_TRANSPORT;
+    return gInfo->IsTransport();
 }
 
 bool GameObject::IsMoTransport() const
